+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### CORE - 0.2.0 - 30.06.24
+* Re-implement `Keypair` that use keys from `rsa` library
+* Implement `Encryptor` and `Decryptor`
+* Re-implement encrypt and decrypt functions using `Encryptor` and `Decryptor` as base
+
 ### CORE - 0.1.0 - 29.06.24
 * Implement Keypair
 * Implement encrypt function